blockchain: encapsulate consensus WIP
