+++ packages/blockchain/test/clique.spec.ts
@@ -3,7 +3,7 @@ import Common, { Chain, ConsensusAlgorithm, ConsensusType, Hardfork } from '@eth
 import { Address } from 'ethereumjs-util'
 import tape from 'tape'
 import Blockchain from '../src'
-import { CLIQUE_NONCE_AUTH, CLIQUE_NONCE_DROP } from '../src/clique'
+import { CLIQUE_NONCE_AUTH, CLIQUE_NONCE_DROP } from '../src/consensus/clique'
 
 tape('Clique: Initialization', (t) => {
   t.test('should initialize a clique blockchain', async (st) => {
@@ -14,7 +14,7 @@ tape('Clique: Initialization', (t) => {
     st.equal(head.hash().toString('hex'), common.genesis().hash.slice(2), 'correct genesis hash')
 
     st.deepEquals(
-      blockchain.cliqueActiveSigners(),
+      (blockchain as any).consensus.cliqueActiveSigners(),
       head.header.cliqueEpochTransitionSigners(),
       'correct genesis signers'
     )
@@ -166,7 +166,7 @@ tape('Clique: Initialization', (t) => {
     }
 
     // calculate difficulty
-    const signers = blockchain.cliqueActiveSigners()
+    const signers = (blockchain as any).consensus.cliqueActiveSigners()
     const signerIndex = signers.findIndex((address: Address) => address.equals(signer.address))
     const inTurn = number % signers.length === signerIndex
     blockData.header.difficulty = inTurn ? BigInt(2) : BigInt(1)
@@ -273,7 +273,7 @@ tape('Clique: Initialization', (t) => {
     const { blocks, blockchain } = await initWithSigners([A])
     const block = await addNextBlock(blockchain, blocks, A)
     st.equal(block.header.number, BigInt(1))
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address])
     st.end()
   })
 
@@ -283,7 +283,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B)
     await addNextBlock(blockchain, blocks, A, [C, true])
     st.deepEqual(
-      blockchain.cliqueActiveSigners(),
+      (blockchain as any).consensus.cliqueActiveSigners(),
       [A.address, B.address],
       'only accept first, second needs 2 votes'
     )
@@ -301,7 +301,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B, [E, true])
 
     st.deepEqual(
-      blockchain.cliqueActiveSigners(),
+      (blockchain as any).consensus.cliqueActiveSigners(),
       [A.address, B.address, C.address, D.address],
       'only accept first two, third needs 3 votes already'
     )
@@ -313,7 +313,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, A, [A, false])
 
     st.deepEqual(
-      blockchain.cliqueActiveSigners(),
+      (blockchain as any).consensus.cliqueActiveSigners(),
       [],
       'weird, but one less cornercase by explicitly allowing this'
     )
@@ -326,7 +326,11 @@ tape('Clique: Initialization', (t) => {
       const { blocks, blockchain } = await initWithSigners([A, B])
       await addNextBlock(blockchain, blocks, A, [B, false])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address], 'not fulfilled')
+      st.deepEqual(
+        (blockchain as any).consensus.cliqueActiveSigners(),
+        [A.address, B.address],
+        'not fulfilled'
+      )
       st.end()
     }
   )
@@ -338,7 +342,7 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, A, [B, false])
       await addNextBlock(blockchain, blocks, B, [B, false])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address], 'fulfilled')
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address], 'fulfilled')
       st.end()
     }
   )
@@ -348,7 +352,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, A, [C, false])
     await addNextBlock(blockchain, blocks, B, [C, false])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
     st.end()
   })
 
@@ -359,7 +363,12 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, A, [C, false])
       await addNextBlock(blockchain, blocks, B, [C, false])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address, C.address, D.address])
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [
+        A.address,
+        B.address,
+        C.address,
+        D.address,
+      ])
       st.end()
     }
   )
@@ -372,7 +381,11 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, B, [D, false])
       await addNextBlock(blockchain, blocks, C, [D, false])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address, C.address])
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [
+        A.address,
+        B.address,
+        C.address,
+      ])
       st.end()
     }
   )
@@ -385,7 +398,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B)
     await addNextBlock(blockchain, blocks, A, [C, true])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
     st.end()
   })
 
@@ -400,7 +413,12 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, A)
     await addNextBlock(blockchain, blocks, B, [C, true])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address, C.address, D.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [
+      A.address,
+      B.address,
+      C.address,
+      D.address,
+    ])
     st.end()
   })
 
@@ -412,7 +430,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B)
     await addNextBlock(blockchain, blocks, A, [B, false])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
     st.end()
   })
 
@@ -430,7 +448,7 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, A)
     await addNextBlock(blockchain, blocks, B, [C, false])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+    st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
     st.end()
   })
 
@@ -441,7 +459,11 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B, [C, false])
     await addNextBlock(blockchain, blocks, A, [B, false])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address], 'deauth votes')
+    st.deepEqual(
+      (blockchain as any).consensus.cliqueActiveSigners(),
+      [A.address, B.address],
+      'deauth votes'
+    )
     st.end()
   })
 
@@ -452,7 +474,11 @@ tape('Clique: Initialization', (t) => {
     await addNextBlock(blockchain, blocks, B, [C, false])
     await addNextBlock(blockchain, blocks, A, [D, true])
 
-    st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address], 'auth votes')
+    st.deepEqual(
+      (blockchain as any).consensus.cliqueActiveSigners(),
+      [A.address, B.address],
+      'auth votes'
+    )
     st.end()
   })
 
@@ -472,7 +498,7 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, A)
       await addNextBlock(blockchain, blocks, C, [C, true])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
       st.end()
     }
   )
@@ -493,7 +519,11 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, A)
       await addNextBlock(blockchain, blocks, B, [C, true])
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address, C.address])
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [
+        A.address,
+        B.address,
+        C.address,
+      ])
       st.end()
     }
   )
@@ -520,7 +550,7 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, D, [A, false])
       await addNextBlock(blockchain, blocks, B, [F, true]) // Finish authorizing F, 3/3 votes needed
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [
         B.address,
         C.address,
         D.address,
@@ -556,7 +586,7 @@ tape('Clique: Initialization', (t) => {
       await addNextBlock(blockchain, blocks, A, undefined, [A, B], common)
       await addNextBlock(blockchain, blocks, B, [C, true], undefined, common)
 
-      st.deepEqual(blockchain.cliqueActiveSigners(), [A.address, B.address])
+      st.deepEqual((blockchain as any).consensus.cliqueActiveSigners(), [A.address, B.address])
       st.end()
     }
   )
@@ -637,24 +667,24 @@ tape('Clique: Initialization', (t) => {
     const { blocks, blockchain } = await initWithSigners([A, B, C])
     // block 1: B, next signer: C
     await addNextBlock(blockchain, blocks, B)
-    st.notOk(await blockchain.cliqueSignerInTurn(A.address))
-    st.notOk(await blockchain.cliqueSignerInTurn(B.address))
-    st.ok(await blockchain.cliqueSignerInTurn(C.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(A.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(B.address))
+    st.ok(await (blockchain as any).consensus.cliqueSignerInTurn(C.address))
     // block 2: C, next signer: A
     await addNextBlock(blockchain, blocks, C)
-    st.ok(await blockchain.cliqueSignerInTurn(A.address))
-    st.notOk(await blockchain.cliqueSignerInTurn(B.address))
-    st.notOk(await blockchain.cliqueSignerInTurn(C.address))
+    st.ok(await (blockchain as any).consensus.cliqueSignerInTurn(A.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(B.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(C.address))
     // block 3: A, next signer: B
     await addNextBlock(blockchain, blocks, A)
-    st.notOk(await blockchain.cliqueSignerInTurn(A.address))
-    st.ok(await blockchain.cliqueSignerInTurn(B.address))
-    st.notOk(await blockchain.cliqueSignerInTurn(C.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(A.address))
+    st.ok(await (blockchain as any).consensus.cliqueSignerInTurn(B.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(C.address))
     // block 4: B, next signer: C
     await addNextBlock(blockchain, blocks, B)
-    st.notOk(await blockchain.cliqueSignerInTurn(A.address))
-    st.notOk(await blockchain.cliqueSignerInTurn(B.address))
-    st.ok(await blockchain.cliqueSignerInTurn(C.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(A.address))
+    st.notOk(await (blockchain as any).consensus.cliqueSignerInTurn(B.address))
+    st.ok(await (blockchain as any).consensus.cliqueSignerInTurn(C.address))
     st.end()
   })
 })
