+++ packages/blockchain/test/reorg.spec.ts
@@ -3,7 +3,7 @@ import { Block } from '@ethereumjs/block'
 import { Address } from 'ethereumjs-util'
 import tape from 'tape'
 import Blockchain from '../src'
-import { CLIQUE_NONCE_AUTH } from '../src/clique'
+import { CLIQUE_NONCE_AUTH } from '../src/consensus/clique'
 import { generateConsecutiveBlock } from './util'
 
 tape('reorg tests', (t) => {
@@ -184,7 +184,7 @@ tape('reorg tests', (t) => {
         'head on the high chain should equal the high block'
       )
 
-      let signerStates = (blockchain as any)._cliqueLatestSignerStates
+      let signerStates = (blockchain as any).consensus._cliqueLatestSignerStates
       t.ok(
         !signerStates.find(
           (s: any) => s[0] === BigInt(2) && s[1].find((a: Address) => a.equals(beneficiary1))
@@ -192,7 +192,7 @@ tape('reorg tests', (t) => {
         'should not find reorged signer state'
       )
 
-      let signerVotes = (blockchain as any)._cliqueLatestVotes
+      let signerVotes = (blockchain as any).consensus._cliqueLatestVotes
       t.ok(
         !signerVotes.find(
           (v: any) =>
@@ -204,7 +204,7 @@ tape('reorg tests', (t) => {
         'should not find reorged clique vote'
       )
 
-      let blockSigners = (blockchain as any)._cliqueLatestBlockSigners
+      let blockSigners = (blockchain as any).consensus._cliqueLatestBlockSigners
       t.ok(
         !blockSigners.find(
           (s: any) => s[0] === BigInt(1) && s[1].equals(block1_low.header.cliqueSigner())
@@ -212,7 +212,7 @@ tape('reorg tests', (t) => {
         'should not find reorged block signer'
       )
 
-      signerStates = (blockchain as any)._cliqueLatestSignerStates
+      signerStates = (blockchain as any).consensus._cliqueLatestSignerStates
       t.ok(
         !!signerStates.find(
           (s: any) => s[0] === BigInt(3) && s[1].find((a: Address) => a.equals(beneficiary2))
@@ -220,10 +220,10 @@ tape('reorg tests', (t) => {
         'should find reorged signer state'
       )
 
-      signerVotes = (blockchain as any)._cliqueLatestVotes
+      signerVotes = (blockchain as any).consensus._cliqueLatestVotes
       t.ok(signerVotes.length === 0, 'votes should be empty')
 
-      blockSigners = (blockchain as any)._cliqueLatestBlockSigners
+      blockSigners = (blockchain as any).consensus._cliqueLatestBlockSigners
       t.ok(
         !!blockSigners.find(
           (s: any) => s[0] === BigInt(3) && s[1].equals(block3_high.header.cliqueSigner())
